Relocated atomic headers to the arch directory
Seems the more appropriate place for them. In addition, I changed the definitions
from function to macro as this will more easily define shortcuts to the GCC primatives.
The reason for this is that the GCC atomic functions are type and size agnostic and doing
without them will require us to write our own size-specific atomic code on the toolchains
that do not support the GCC atomics (looking at you IBM...). Anyways, no need to define a
full wrapper function when we can do a 1-line macro.