Partial Move of BG Linux Tree revieve functionality
Since we are building on the IBM toolchain, we now have access to the
bg-specific PPC450 double-hummer instuctions. I've replaced our old macro
'hack' with the actual instructions.

I've made a partial move of the BG linux recive procedure code into our TREE
Ebb. It compiles but is nowhere near ready for a test run.